fetch data

+++ app/src/main/java/com/sardordev/budgetapp/data/model/AmountData.kt
@@ -5,5 +5,5 @@ import androidx.room.ColumnInfo
 
 data class AmountData(
     @ColumnInfo(name = "amountbudget")
-    val amountBudgets: Int
+    val amountBudgets: Float
 )
